SLING-9077: implemented an optimised PathSet with O(n+m) additionally to O(n^2)
There are various use cases for PathSets, most of them involve small sets of
paths or glob patterns. In those cases a polynominal time complexity for
optimising the PathSet does not harm performance noticeable. Other use cases
on the other hand require large sets of only paths for which a linear time
complexity can be achieved. Here a compromise between the both use  cases was
implemented. For small sets the simple but expensive optimisation is applied
and once the size of the PathSet exceeds the theoretical bound of the simple
algorithm it switches to the less expensive implementation.