SLING-11878 | Update ModifiableValueMap javadoc to cover missing methods inherited from java.util.Map
The original javadoc for `ModifiableValueMap` was written before the
addition of new methods to `java.util.Map` as of Java 8. This commit
updates the javadoc to mention the following methods that can can be
used to modify resources underlying a `ModifiableValueMap`:

- `replace`
- `replaceAll`
- `compute`
- `computeIfAbsent`
- `computeIfPresent`
- `merge`

Refer to https://issues.apache.org/jira/browse/SLING-11878 for context.